RAG/requirements.txt
flask
python-dotenv
langchain
langchain-core
langchain-community
numpy
pandas
scikit-learn
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-google-genai
google-generativeai
flask
python-dotenv